test(#2429): batch 12 — migrate I-group merge/gateway/org-sync/run-as cluster (8 dirs, 336 items) into canonical layers
- 822 merge-conflict detection (75), 1001 API-error retry (25), 584 remote
  create-directory (24) -> tests/unit; 584 sys.modules/sys.path leaks scoped
  via patch.dict + spec loader
- 720 model-gateway: planner (25) + handler (22) + admin view/security classes
  (29) -> tests/unit; repository classes (5, real tmp sqlite + _DDL kept) ->
  tests/integration; SSRF service tests fail-safed with safe_request mock
- 1891 usage-report auth: HTTP/auth/binding/audit bulk (28) ->
  tests/integration/routes; input validation (9) -> tests/unit; dead ambient-DB
  session_manager fixture dropped
- 1829: DingTalk token cache (5) + fetch_openclaw source (8, raw env/reload/
  sys.path replaced by scoped loader) + tool-account route (5) -> tests/unit;
  signing-key multitenancy (12, +security) + tool-account repo (3) ->
  tests/integration
- 1827 org-sync hardening (27) -> tests/unit, REPAIRED: red on main since
  #3028/#3029 (fakes never realigned to snapshot 3-tuple + kwargs +
  (users, complete) return); realigned with strengthened assertions plus the
  collaboration.is_postgresql sqlite pin
- 2018 run-as gate: validator (22) -> tests/unit; bash gate (8, +security)
  and Linux-root integration (4, skipif verbatim) ->
  tests/integration/subprocess

Baselines: issues 1856->1520 / 207->192 (exact -336/-15); default min
6233->6569 (+336); ci floor >=192; inventory 108->100 entries; 4 empty
__init__.py removed. Scanner: 0 new findings (69 = main).
