+++ .test-baseline.json
@@ -4,12 +4,12 @@
   "created_before_modification": true,
   "layers": {
     "default": {
-      "min_tests": 6233,
+      "min_tests": 6569,
       "min_files": 250,
-      "last_updated": "2026-08-24",
-      "update_reason": "Issue #2429 batch 11: subprocess-core cluster migrated 273 items (251 unit + 22 integration); min raised by exactly +273 (5960+273); actual 9432 = 9159 (main pre-batch) + 273",
-      "actual_collected": 9432,
-      "actual_files": 653,
+      "last_updated": "2026-08-27",
+      "update_reason": "Issue #2429 batch 12: merge/model-gateway/org-sync/run-as cluster migrated 336 items (unit: 822/1001/584/720 planner+handler+admin view-classes/1891 validation/1829 cache+fetch+tool-account-route/1827/2018 validator; integration: 720 repository, 1891 routes, 1829 signing+tool-account-repo, 2018 gate+integration); min raised by exactly +336 (6233+336); actuals 10203/690 measured locally on branch (= 9867 base 6bdeebea + 336; CI clean re-confirmed)",
+      "actual_collected": 10203,
+      "actual_files": 690,
       "threshold_percent": 96
     },
     "critical": {
@@ -30,12 +30,12 @@
       "note": "Currently excluded by norecursedirs"
     },
     "issues": {
-      "min_tests": 1856,
-      "min_files": 207,
-      "last_updated": "2026-08-24",
-      "update_reason": "Issue #2429 batch 11: I-group subprocess-core cluster (8 dirs: 559/1395/1002/786/1005/1461/2339/2340, 273 items / 14 glob files) — all 273 migrated (251 -> tests/unit incl. dedup filter swapped to the real production predicate, 22 -> tests/integration: subprocess/ for the 559 loopback-WS self-subprocess script [renamed off the e2e_ prefix] and the 1395 wrapper cluster, concurrency/ for 1005 pause-budget wall-clock and the 1002 heartbeat-thread cleanup half); 4 empty __init__.py deleted; 559 e2e's two vacuous rejection checks given real outcome assertions; 1005 wait-routing test hermeticized against build_secure_agent_env; min lowered by exact deltas (2129-273, 221-14)",
-      "actual_collected": 2068,
-      "actual_files": 222,
+      "min_tests": 1520,
+      "min_files": 192,
+      "last_updated": "2026-08-27",
+      "update_reason": "Issue #2429 batch 12: I-group merge/model-gateway/org-sync/run-as cluster (8 dirs: 822/720/1891/1829/1827/2018/1001/584, 336 items / 15 glob files) — all 336 migrated to canonical layers (213 unit incl. repaired 1827, 123 integration: 720 repository + 1891 usage-report routes + 1829 signing/tool-account-repo + 2018 run-as gate [security] & root-integration [skipif kept]); 4 empty __init__.py deleted; 1827 was red on main since #3028/#3029 (fakes never realigned: snapshot 3-tuple + kwargs + (users, complete) return) — repaired with strengthened assertions plus the collaboration.is_postgresql sqlite pin; 1891 dead ambient-DB fixture dropped; 1829 fetch raw env/reload/sys.path replaced by scoped loader; 584 sys.modules/sys.path leaks scoped; 720 admin SSRF service tests fail-safed with safe_request mock; min lowered by exact deltas (1856-336, 207-15)",
+      "actual_collected": 1733,
+      "actual_files": 207,
       "threshold_percent": 97
     }
   },
@@ -47,7 +47,7 @@
     "Baseline created before pytest.ini modification to enable e2e_*.py collection",
     "tests/e2e currently excluded by norecursedirs in pytest.ini",
     "Issue item baseline is enforced by ci.yml; the extended runner scales the file baseline per shard",
-    "Default count from: python scripts/ci.py run default-collection (9432 items after #2429 batch 11)",
-    "Issue count from: clean-checkout issue-collection (2341 items, 238 files after #2429 batch 10; +1/+1 local gitignored tests/issues/89 noise)"
+    "Default count from: python scripts/ci.py run default-collection (10203 items after #2429 batch 12)",
+    "Issue count from: clean-checkout issue-collection (1733 items, 207 files after #2429 batch 12; base 6bdeebea was 2069/222 — main gained +1 issue item after batch 11 via 5fe363be; no tests/issues/89 in clean checkouts)"
   ]
 }
