+++ tests/issues/legacy-directories.txt
@@ -63,23 +63,19 @@
 435
 517
 555
-584
 604
 610
 679
 716
-720
 740
 761
 776
-822
 824
 863
 883
 886
 913
 993
-1001
 1006
 1138
 1140
@@ -92,16 +88,12 @@
 1751
 1771
 1776
-1827
-1829
 1852
 1854
 1855
 1857
 1858
-1891
 1892
-2018
 2019
 2020
 2021
